Reorganized source folder
Moved all source code to private folder.
Added flags for toggling the generation of suggestions on/of via
BIPlugin_GenerateSuggestions
Added logs of timings for various activities within the generation of
suggestions in order to identify bottlenecks.
More elaborate presentation of results of the K-Fold cross validation.
Nodes are shuffled in a predictive manner before use in order to get
more stable results.